refact: move HTTP client to its own file
